pyproject.toml
[project]
name = "django_id_areadata"
version = "0.1.0"
description = "Django App to Get Indonesian Area Data"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    { name = "Hendika N.", email = "hendika.nugroho@visiprima.id" }
]
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.6",
    "django-filter>=25.1",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
